// webapp/agnor-backend/src/agnoruser/interface/agnoruser.interface.ts
export interface AgnoruserI {
    id?: number;
    usuario?: string;
    correo?: string;
    nombre?: string;
    apellidos?: string;
    edad?: number;
    telefono?: number;
    contrasena?: string;
    ultimasesion?: number;
    sesionfecha?: number;
    grupo?: number;
    coordinador?: string;
    uinmuebles24?: string;
    cinmuebles24?: string;
    uvivanuncios?: string;
    cvivanuncios?: string;
    rol?:string;
}